sun morn

+++ src/java/AssayRunViewer.java
@@ -74,8 +74,9 @@ public class AssayRunViewer extends JDialog implements java.awt.event.ActionList
 
     //    this.session = dmf.getSession();
      IFn getProjectID = Clojure.var("ln.codax-manager", "get-project-id");
-
-     project_id = ((Integer)getProjectID.invoke()).intValue();
+     //cannot be (Integer)    ...intValue()
+     // project is long, others are int??
+     project_id = ((Long)getProjectID.invoke()).intValue();
       IFn getUser = Clojure.var("ln.codax-manager", "get-user");
    
     owner = (String)getUser.invoke();
@@ -89,7 +90,8 @@ public class AssayRunViewer extends JDialog implements java.awt.event.ActionList
     assay_runs_pane.setBorder(assay_runs_pane_border);
 
     //this failed as Long
-    assay_runs_table = dbm.getDatabaseRetriever().getAssayRuns(((int)getProjectID.invoke()));
+    //getAssayRuns requires and int
+    assay_runs_table = dbm.getDatabaseRetriever().getAssayRuns(((Long)getProjectID.invoke()).intValue());
   assay_runs_table.getSelectionModel().addListSelectionListener(						     
 	  new ListSelectionListener() {
 	      public void valueChanged(ListSelectionEvent e) {
@@ -147,7 +149,8 @@ public class AssayRunViewer extends JDialog implements java.awt.event.ActionList
     hit_lists_pane_border.setTitlePosition(javax.swing.border.TitledBorder.TOP);
     hit_lists_pane.setBorder(hit_lists_pane_border);
     //failed as Long
-    hit_lists_table = dbm.getDatabaseRetriever().getHitLists(((int)getProjectID.invoke()));
+    // getHitLists requires int
+    hit_lists_table = dbm.getDatabaseRetriever().getHitLists(((Long)getProjectID.invoke()).intValue());
 
     hit_lists_scroll_pane = new JScrollPane(hit_lists_table);
     hit_lists_table.setFillsViewportHeight(true);
